# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JobMatch API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all scraped jobs
      parameters:
        - name: jobType
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: search
          in: query
          description: Search across all searchable fields of each job
          schema:
            type: string
        - name: category
          in: query
          description: Filter jobs by their shared job classification
          schema:
            type: string
            enum:
              - Finance & Comptabilité
              - Commercial & Vente
              - Administration
              - Marketing & Communication
              - Informatique & Tech
              - Ressources humaines
              - Logistique & Transport
              - Industrie & Technique
              - Santé
              - Éducation
              - BTP & Immobilier
              - Hôtellerie & Restauration
              - Agriculture
              - Juridique
              - Autres
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          description: Number of jobs to skip before returning results
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
  /public/jobs/{id}:
    get:
      operationId: getPublicJobSharePreview
      tags: [jobs]
      summary: Get the public teaser for a shared job
      description: Returns only the non-Premium information safe to display on a public sharing page.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Public job teaser
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicJobSharePreview"
        "404":
          description: Job not found or no longer available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /jobs/scrape:
    post:
      operationId: triggerScrape
      tags: [jobs]
      summary: Trigger job scraping
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScrapeRequest"
      responses:
        "200":
          description: Scraping result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrapeResult"
...
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a single job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SessionToken
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
...
    Job:
      type: object
      required: [id, title, company, description, location, skills, jobType, source, scrapedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        logoUrl:
          type: ["string", "null"]
          description: URL du logo de l'entreprise
        description:
          type: string
        location:
          type: string
        remote:
          type: boolean
        skills:
          type: array
          items:
            type: string
        jobType:
          type: string
        salaryMin:
          type: ["integer", "null"]
        salaryMax:
          type: ["integer", "null"]
        source:
          type: string
        sourceUrl:
          type: ["string", "null"]
        postedAt:
          type: ["string", "null"]
        experienceYears:
          type: ["integer", "null"]
        sector:
          type: ["string", "null"]
        diplomaRequired:
          type: ["string", "null"]
        scrapedAt:
          type: string
    PublicJobSharePreview:
      type: object
      required: [id, title, company, location, remote, skills, jobType, source, postedAt, scrapedAt, summary, isVerified]
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        logoUrl:
          type: ["string", "null"]
        location:
          type: string
        remote:
          type: boolean
        skills:
          type: array
          items:
            type: string
        jobType:
          type: string
        source:
          type: string
        postedAt:
          type: ["string", "null"]
        scrapedAt:
          type: string
        sector:
          type: ["string", "null"]
        experienceYears:
          type: ["integer", "null"]
        summary:
          type: string
        isVerified:
          type: boolean
...
    ScrapeRequest:
      type: object
      properties:
        keywords:
          type: string
        location:
          type: string
    ScrapeResult:
      type: object
      required: [jobsAdded, jobsUpdated, total]
      properties:
        jobsAdded:
          type: integer
        jobsUpdated:
          type: integer
        total:
          type: integer
        message:
          type: string
...